Add possibility of running project with arguments
The only way to pass parameters from the command line (to java) seem to
be properties, according to [1].

However if an "arg" element is inside a "java" task, the java program
always receives an argument, even if the user doesn't set the property.
To fix this the if-unless pattern [2] of ant was used.

This way of passing parameters to the program may have the limitation of
allowing only one parameter to be passed, even if given a "string"
between quotation marks.

[1] - http://ant.apache.org/faq.html#passing-cli-args
[2] - http://stackoverflow.com/a/14501217